Rename initializer dir to init.
