+++ src/cli/commands/init/initTokenHandler.js
